GUARDIA--AI-Powered-Entry-Gate-Robo-master/requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.7
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
dlib==19.24.6
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.9
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
jiter==0.7.0
kiwisolver==1.4.7
MarkupSafe==2.1.5
numpy==1.26.4
openai==1.53.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
protobuf==4.25.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pyttsx3==2.91
requests==2.32.3
sentencepiece==0.2.0
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
SpeechRecognition==3.10.4
tqdm==4.66.6
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
wheel==0.44.0

azure-cognitiveservices-speech==1.41.1
mysql-connector-python==9.1.0